+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] / 2022-07-09
+### Added
+- Add `CreatePushButton` and Obsolete `AddPushButton`
+### Updated
+- Update `RibbonSettings` Add with Action
+- Update `RibbonDescription` (LanguageType) 
+
 ## [0.1.2] / 2022-05-17
 ### Added
 - Add `IExternalCommandAvailability` on `NewPushButtonData` Command
@@ -151,6 +158,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - First Release
 
 [vNext]: ../../compare/1.0.0...HEAD
+[0.2.0]: ../../compare/0.1.2...0.2.0
+[0.1.2]: ../../compare/0.1.1...0.1.2
 [0.1.1]: ../../compare/0.1.0...0.1.1
 [0.1.0]: ../../compare/0.0.9...0.1.0
 [0.0.9]: ../../compare/0.0.8...0.0.9